requirements.txt
bcrypt==4.2.0
blinker==1.8.2
click==8.1.7
Deprecated==1.2.18
Flask==3.0.3
Flask-Limiter==3.10.1
Flask-WTF==1.2.2
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
limits==4.0.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mysql-connector-python==9.1.0
numpy==2.2.3
ordered-set==4.1.0
packaging==24.2
pandas==2.2.3
Pygments==2.19.1
python-dateutil==2.9.0.post0
pytz==2025.1
rich==13.9.4
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2025.1
Werkzeug==3.1.1
wrapt==1.17.2
WTForms==3.2.1
